feat: add about button to options page
This commit adds an 'About' button to the options page. The button is styled with CSS and includes an event listener that opens an 'about.html' page in a new window when clicked. The button is positioned at the bottom right of the page and scales slightly when hovered over.